backend/requirements.txt
Django=5.2
djangorestframework=3.16.0
Pillow=11.2.1
djoser=2.3.1
django-filter=25.1
requests=2.32.3